+++ packages/log-http/middleware/not-found.js
@@ -1,0 +1,4 @@
+'use strict';
+
+module.exports = () => (req, res) =>
+    res.status(404).end(`Cannot ${req.method} ${req.url}`);